requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-bigquery==3.20.1
google-cloud-secret-manager==2.20.0
google-auth==2.29.0
google-api-python-client==2.130.0
google-cloud-storage==2.16.0
vertexai==1.56.0
requests==2.31.0
